refactor: 3단계 피드백 반영 - 일급컬렉션 적용
