+++ src/test/java/bowling/domain/state/StateFactoryTest.java
@@ -1,7 +1,5 @@
 package bowling.domain.state;
 
-import bowling.domain.frame.Frame;
-import bowling.domain.frame.NormalFrame;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,15 +9,11 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class StateFactoryTest {
 
-    private Frame finalFrame;
+    private States states;
 
     @BeforeEach
-    void setup() throws Exception {
-        finalFrame = NormalFrame.initialize();
-        for (int i = 1; i < 10; i++) {
-            finalFrame.bowling(10);
-            finalFrame = finalFrame.next();
-        }
+    void setup() {
+        states = States.initialize();
     }
 
     @Test
@@ -31,34 +25,35 @@ class StateFactoryTest {
     @Test
     @DisplayName("스트라이크 두번 시 다음 State 를 BonusReady 를 반환한다.")
     void nextState_BonusReady() {
-        finalFrame.bowling(10);
-        finalFrame.bowling(10);
+        states.bowl(10);
+        states.bowl(10);
 
-        assertThat(StateFactory.nextState(finalFrame)).isInstanceOf(BonusReady.class);
+        assertThat(StateFactory.nextState(states)).isInstanceOf(BonusReady.class);
     }
 
     @Test
     @DisplayName("스페어 처리 시 다음 State 를 BonusReady 를 반환한다.")
     void nextState_BonusReady_2() {
-        finalFrame.bowling(3);
-        finalFrame.bowling(7);
+        states.bowl(3);
+        states.bowl(7);
 
-        assertThat(StateFactory.nextState(finalFrame)).isInstanceOf(BonusReady.class);
+        assertThat(StateFactory.nextState(states)).isInstanceOf(BonusReady.class);
     }
 
     @Test
     @DisplayName("스트라이크 한번 했을 때 다음 State 를 Ready 를 반환한다.")
     void nextState_Ready() {
-        finalFrame.bowling(10);
+        states.bowl(10);
 
-        assertThat(StateFactory.nextState(finalFrame)).isInstanceOf(Ready.class);
+        assertThat(StateFactory.nextState(states)).isInstanceOf(Ready.class);
     }
 
     @Test
     @DisplayName("State 가 종료되지 않았을 때 다음 State 를 호출하면 예외를 반환한다.")
     void invalidNextState() {
-        finalFrame.bowling(4);
+        states.bowl(4);
 
-        assertThatThrownBy(() -> StateFactory.nextState(finalFrame)).isInstanceOf(IllegalStateException.class);
+        assertThat(states.isFinish()).isFalse();
+        assertThatThrownBy(() -> StateFactory.nextState(states)).isInstanceOf(IllegalStateException.class);
     }
 }
